Better index.js building & Fix typo
